requirements.txt
requests~=2.32.4
google-cloud-storage~=3.1.1